LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.put.clip.worker-expected.txt


PASS putImageData() is not affected by clipping regions
